Add files via upload
This file contains cleaned and preprocessed movie data from The Movie Database (TMDB), spanning 2010-2023, with a focus on films produced in the UK and US. The accompanying Jupyter Notebook provides a detailed account of the data cleaning and preprocessing steps, including decisions and reasoning. This dataset serves as a foundation for analysis and visualization of trends and patterns in the US and UK movie industries and the factors affecting Profit making in the menioned movie industries.